pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sddraft"
version = "0.1.0"
description = "Generate and update MIL-STD-498-style SDD drafts from repository evidence."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "SDDraft Contributors" }
]
dependencies = [
  "pydantic>=2.8,<3.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "black==24.10.0",
  "isort",
  "ruff",
  "mypy",
  "types-PyYAML",
  "pytest",
  "pytest-cov",
]

[project.scripts]
sddraft = "sddraft.cli.main:main"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.ruff_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["sddraft"]
skip = [".venv", ".mypy_cache", ".ruff_cache", "build", "dist"]

[tool.ruff]
line-length = 88
target-version = "py311"
respect-gitignore = true
src = ["src"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "I",
  "B",
  "UP",
]
ignore = [
  "E501",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
packages = ["sddraft"]
mypy_path = ["src"]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra --cov=sddraft --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["sddraft"]
branch = true
omit = [
  "src/sddraft/__main__.py",
]
disable_warnings = [
  "module-not-measured",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 80

[tool.setuptools.packages.find]
where = ["src"]
include = ["sddraft*"]
